a more complex test (after the fact); parsed operation should return a floating point value if possible

+++ src/parse.py
@@ -4,7 +4,7 @@ def parse_operation(operation_string):
   arithmetic = re.compile('^[+\-*/][0-9]+$')
   if (arithmetic.match(operation_string)):
     operator = operation_string[0:1]
-    second_term = int(operation_string[1:])
+    second_term = float(operation_string[1:])
 
     return {
       '+': lambda value: value + second_term,
